Automation_Selenium/Different_Techniques/refactored_pycharm_test/requirements.txt
selenium
pytest
pytest-html
webdriver-manager
pyyaml
python-dotenv
